Move apps tests to right folder
